Upgrade to Neo4j 4.0.0 driver.
This commit has 3 aspects, the upgrade to the Neo4j 4.0.0 driver being
the primary concern.

For the upgrade package imports needed to be fixed and one of the
builder methods of the Neo4j bolt configuration needed a small change.

As we distinguish now between three session types, those sessions live
in different packages as well.

As the asynchronous session has been used for the Health check, that
needed to be fixed as well.

The health indicator is the second aspect. The composition of the
query and results can be simplified. Also we should use a write
transaction, in case we are in a cluster scenario. A read transaction
would only indicate general health but a status of `UP` would not
guarantee writes to succede.
At the same position, we strongly recommends against counting nodes for
health and propose returning some info about the server itself.

Last but not least, the driver has build-in slf4 logging, so that the
anonymous inner class can just be dropped.